Feat: delete player confirmation

+++ src/main/java/com/nathan/footballsquadmanagerbp2/controller/AllPlayersController.java
@@ -1,9 +1,14 @@
 package com.nathan.footballsquadmanagerbp2.controller;
 
 import com.nathan.footballsquadmanagerbp2.model.Player;
+import com.nathan.footballsquadmanagerbp2.service.AlertService;
 import com.nathan.footballsquadmanagerbp2.service.PlayerService;
 import com.nathan.footballsquadmanagerbp2.view.PlayerDetailsView;
 import javafx.collections.ObservableList;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
+
+import java.util.Optional;
 
 public class AllPlayersController {
     PlayerDetailsView showPlayerView;
@@ -18,7 +23,16 @@ public class AllPlayersController {
     }
 
     public void deletePlayer(Player player) {
-        playerService.deletePlayer(player);
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Delete Player");
+        alert.setContentText("Are you sure you want to delete this player?");
+        alert.setHeaderText(null);
+
+        Optional<ButtonType> result = alert.showAndWait();
+
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            playerService.deletePlayer(player);
+        }
     }
 
     public ObservableList<Player> getAllPlayers() {
